Move cwp graph building over to the builder interface
